+++ src/itest/scala/conf/play.plugins
