+++ src/main/scala/play/modules/reactivemongo/ReactiveMongoPluginException.scala
@@ -1,0 +1,3 @@
+package play.modules.reactivemongo
+
+class ReactiveMongoPluginException(message: String, cause: Throwable = null) extends RuntimeException(message, cause)
