+++ src/main/scala/play/modules/reactivemongo/json/Writers.scala
@@ -1,0 +1,26 @@
+package play.modules.reactivemongo.json
+
+import play.api.libs.json._
+
+object Writers {
+  implicit class JsPathMongo(val jp: JsPath) extends AnyVal {
+    def writemongo[A](implicit writer: Writes[A]): OWrites[A] = {
+      OWrites[A] { (o: A) =>
+        val newPath = jp.path.flatMap {
+          case e: KeyPathNode     => Some(e.key)
+          case e: RecursiveSearch => Some(s"$$.${e.key}")
+          case e: IdxPathNode     => Some(s"${e.idx}")
+        }.mkString(".")
+
+        val orig = writer.writes(o)
+        orig match {
+          case JsObject(e) =>
+            JsObject(e.flatMap {
+              case (k, v) => Seq(s"$newPath.$k" -> v)
+            })
+          case e: JsValue => JsObject(Seq(newPath -> e))
+        }
+      }
+    }
+  }
+}
